feat: Abstract key command.
