+++ src/node/LibplanetConsole.NodeHost/Application.cs
@@ -2,7 +2,6 @@ using JSSoft.Commands.Extensions;
 using JSSoft.Terminals;
 using LibplanetConsole.Common.Extensions;
 using LibplanetConsole.Nodes;
-using Serilog;
 
 namespace LibplanetConsole.NodeHost;
